Splitting JavaScript into different files
Stop us from FUCKING EACH OTHER
