+++ BucsEyeView/index.html
@@ -19,16 +19,14 @@ The last section connects the webpage to an external javascript file.
     <link rel="stylesheet" href="css/styles.css">
 </head>
 <body>
+
     <!-- Game Screen -->
     <div id="gameScreen">
         <canvas id="gameCanvas" width="800" height="400"></canvas>
     </div>
-
-    <!-- Added missing startGameButton -->
-    <div id="startGameButton" style="display: none;">
-        <!-- Content for the button -->
-    </div>
-
+    
+    <script src="js/menu.js"></script>
+    <script src="js/gamelogic.js"></script>
     <script src="js/main.js"></script>
 </body>
 </html>
